pyproject.toml
[tool.poetry]
name = "lexguard-contract-ai"
version = "0.1.0"
description = "AI-powered legal contract analyzer with risk scoring, clause classification, and interactive Q&A"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "lexguard"}, {include = "backend"}, {include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pandas = "^2.1.3"
python-multipart = "^0.0.6"
pypdf = "^3.17.1"
pytesseract = "^0.3.10"
pillow = "^10.1.0"
tiktoken = "^0.5.1"
chromadb = "^0.4.18"
sentence-transformers = "^2.2.2"
openai = "^1.3.7"
python-dotenv = "^1.0.0"
reportlab = "^4.0.7"
streamlit = "^1.28.2"
httpx = "^0.25.2"
numpy = "^1.26.2"
plotly = "^6.4.0"
python-docx = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.6"
black = "^23.11.0"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


